Add command `cover` to the dzil app
The `cover` builds and prepares the distribution. Whne tests are ready
to run, it uses the Devel::Cover command line tool `cover`, with a
`-test` parameter to run the distribution test suite generating coverage
information.

The `cover_db/` database is rsync'ed to the main distribution directory.

If the optional parameter `-o` is given, it opens the
`cover_db/coverage.html` file.

Signed-off-by: Pedro Melo <melo@simplicidade.org>